gs-backend/requirements.txt
# store
aiomysql
sqlalchemy~=1.4.31
PyMySQL~=1.0.2

# utils
markdown~=3.3.6
Pygments~=2.11.2
pytz~=2021.3
jinja2~=3.1.2
psutil~=5.9.3

# logic
pyzmq~=22.3.0
pyopenssl~=23.0.0

# api
websockets==10.4
sanic==21.12.2
sanic-ext==21.12.3
sanic-routing==0.7.2
httpx[http2]~=0.23.0

# admin
flask~=2.0.2
flask-sqlalchemy~=2.5.1
flask-admin~=1.6.0
Flask-Migrate~=3.1.0
wtforms~=3.0.1
markupsafe~=2.1.1
gevent~=21.12.0

objprint
pitricks==0.1.3
aiosmtplib